+++ src/test/java/chess/dao/spark/DatabaseConnectorTest.java
@@ -1,4 +1,4 @@
-package chess.dao;
+package chess.dao.spark;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatNoException;
